+++ lib/engineyard/cli.rb
@@ -69,8 +69,8 @@ module EY
       :desc => "Name of the account in which the environment can be found"
     method_option :verbose, :type => :boolean, :aliases => %w(-v),
       :desc => "Be verbose"
-    method_option :extra_deploy_hook_options, :type => :hash, :default => {},
-      :desc => "Additional options to be made available in deploy hooks (in the 'config' hash)"
+    method_option :config, :type => :hash, :default => {}, :aliases => %w(--extra-deploy-hook-options),
+      :desc => "Hash made available in deploy hooks (in the 'config' hash), can also override some ey.yml settings."
     def deploy
       app_env = fetch_app_environment(options[:app], options[:environment], options[:account])
 
@@ -258,8 +258,8 @@ module EY
       :desc => "Name of the account in which the environment can be found"
     method_option :verbose, :type => :boolean, :aliases => %w(-v),
       :desc => "Be verbose"
-    method_option :extra_deploy_hook_options, :type => :hash, :default => {},
-      :desc => "Additional options to be made available in deploy hooks (in the 'config' hash)"
+    method_option :config, :type => :hash, :default => {}, :aliases => %w(--extra-deploy-hook-options),
+      :desc => "Hash made available in deploy hooks (in the 'config' hash), can also override some ey.yml settings."
     def rollback
       app_env = fetch_app_environment(options[:app], options[:environment], options[:account])
       env_config    = config.environment_config(app_env.environment_name)