+++ spec/ey/deploy_spec.rb
@@ -354,12 +354,12 @@ describe "ey deploy" do
     end
   end
 
-  context "--extra-deploy-hook-options" do
+  context "--config (--extra-deploy-hook-options)" do
     before(:all) do
       login_scenario "one app, one environment"
     end
 
-    def extra_deploy_hook_options
+    def config_options
       if @ssh_commands.last =~ /--config (.*?)(?: -|$)/
         # the echo strips off the layer of shell escaping, leaving us
         # with pristine JSON
@@ -367,11 +367,18 @@ describe "ey deploy" do
       end
     end
 
-    it "passes the extra configuration to engineyard-serverside" do
+    it "passes --config to engineyard-serverside" do
+      ey %w[deploy --config some:stuff more:crap --no-migrate]
+      config_options.should_not be_nil
+      config_options['some'].should == 'stuff'
+      config_options['more'].should == 'crap'
+    end
+
+    it "supports legacy --extra-deploy-hook-options" do
       ey %w[deploy --extra-deploy-hook-options some:stuff more:crap --no-migrate]
-      extra_deploy_hook_options.should_not be_nil
-      extra_deploy_hook_options['some'].should == 'stuff'
-      extra_deploy_hook_options['more'].should == 'crap'
+      config_options.should_not be_nil
+      config_options['some'].should == 'stuff'
+      config_options['more'].should == 'crap'
     end
 
     context "when ey.yml is present" do
@@ -382,8 +389,8 @@ describe "ey deploy" do
       after { File.unlink("ey.yml") }
 
       it "overrides what's in ey.yml" do
-        fast_ey %w[deploy --extra-deploy-hook-options beer:esb]
-        extra_deploy_hook_options['beer'].should == 'esb'
+        fast_ey %w[deploy --config beer:esb]
+        config_options['beer'].should == 'esb'
       end
     end
   end