+++ spec/ey/rollback_spec.rb
@@ -27,12 +27,12 @@ describe "ey rollback" do
     @ssh_commands.last.should =~ /--stack nginx_mongrel/
   end
 
-  context "--extra-deploy-hook-options" do
+  context "--config (--extra-deploy-hook-options)" do
     before(:all) do
       login_scenario "one app, one environment"
     end
 
-    def extra_deploy_hook_options
+    def config_options
       if @ssh_commands.last =~ /--config (.*?)(?: -|$)/
         # the echo strips off the layer of shell escaping, leaving us
         # with pristine JSON
@@ -40,11 +40,18 @@ describe "ey rollback" do
       end
     end
 
-    it "passes the extra configuration to engineyard-serverside" do
+    it "passes --config to engineyard-serverside" do
+      ey %w[rollback --config some:stuff more:crap]
+      config_options.should_not be_nil
+      config_options['some'].should == 'stuff'
+      config_options['more'].should == 'crap'
+    end
+
+    it "supports legacy --extra-deploy-hook-options" do
       ey %w[rollback --extra-deploy-hook-options some:stuff more:crap]
-      extra_deploy_hook_options.should_not be_nil
-      extra_deploy_hook_options['some'].should == 'stuff'
-      extra_deploy_hook_options['more'].should == 'crap'
+      config_options.should_not be_nil
+      config_options['some'].should == 'stuff'
+      config_options['more'].should == 'crap'
     end
 
     context "when ey.yml is present" do
@@ -55,8 +62,8 @@ describe "ey rollback" do
       after { File.unlink("ey.yml") }
 
       it "overrides what's in ey.yml" do
-        ey %w[rollback --extra-deploy-hook-options beer:esb]
-        extra_deploy_hook_options['beer'].should == 'esb'
+        ey %w[rollback --config beer:esb]
+        config_options['beer'].should == 'esb'
       end
     end
   end